Allow `root` user operation in `init-zebra-env` service.
This commit fixes a new permissions error when attempting to bring up
the zebrad service for the first time on a host whose volume does not
have the `/var/cache/zebrad-cache` directory already created and set to
the appropriate permissions. The issue is casued by the `latest`
upstream container's use of `gosu` to drop root privileges as part of
their `entrypoint.sh` script, which reverts the `user: "root"` directive
in the Docker Compose file to run as their non-privileged `zebra` user.