The `zebra` service fails to report healthy due to no `curl` installed.
The `healthcheck` in the `zebra` service in these Docker Compose files
is relying on the presence of `curl`, which the container image does not
ship with. This casues the health check to reliably fail and means that
any dependent services also fail to start.

This commit fixes the issue by not relying on `curl` and instead reading
the raw value from the `/proc/net/tcp` state out of the kernel to test
for the presence of an active socket (of any state) at the RPC port. The
RPC port is 8232 in decimal, but is `printf`'ed in hexadecimal to match
the output of the `/proc` filesystem's `net/tcp` output.

We also add some more documentation to describe what a healthy (syncing)
output from the `zebra` service looks like and how to access its
JSON-RPC server with an example command, the `getpeerinfo` method.