Update to modern Docker Compose defaults.
This commit is mostly documentation and convention updates to conform
more closely to the latest Docker Compose specification. It renames
files to the modern Compose defaults, adds `.gitignore` lines that clean
up some cruft in common Git command output, and more explicitly defines
port publication addresses to remind newbies that we are binding to all
(and public) interfaces in the majority of cases.